requirements.txt
mock
asynctest
pyyaml
aiohttp
sqlalchemy
trafaret
trafaret_validator